+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 
 ### Feature
 
+- Add NumberWidget, an input widget for numbers @tiberiuichim
+
 ### Bugfix
 
 ### Internal